identity-lab/requirements.txt
fastapi>=0.100
uvicorn>=0.23
python-jose>=3.3
requests>=2.31
